add lv_mem_dcache
