+++ src/misc/lv_mem.c
@@ -13,7 +13,6 @@
 #include "lv_log.h"
 #include LV_STDLIB_INCLUDE
 #include LV_STRING_INCLUDE
-#include LV_DCACHE_INCLUDE
 
 #if LV_USE_BUILTIN_MALLOC
     #include "lv_malloc_builtin.h"
@@ -181,42 +180,6 @@ void lv_mem_monitor(lv_mem_monitor_t * mon_p)
 #if LV_USE_BUILTIN_MALLOC
     lv_mem_monitor_builtin(mon_p);
 #endif
-}
-
-void lv_mem_clean_dcache_all(void)
-{
-    LV_MEM_CLEAN_DCACHE_ALL;
-}
-
-void lv_mem_clean_dcache(lv_uintptr_t start, lv_uintptr_t end)
-{
-    LV_UNUSED(start);
-    LV_UNUSED(end);
-    LV_MEM_CLEAN_DCACHE;
-}
-
-void lv_mem_flush_dcache_all(void)
-{
-    LV_MEM_FLUSH_DCACHE_ALL;
-}
-
-void lv_mem_flush_dcache(lv_uintptr_t start, lv_uintptr_t end)
-{
-    LV_UNUSED(start);
-    LV_UNUSED(end);
-    LV_MEM_FLUSH_DCACHE;
-}
-
-void lv_mem_invalidate_dcache_all(void)
-{
-    LV_MEM_INVALIDATE_DCACHE_ALL;
-}
-
-void lv_mem_invalidate_dcache(lv_uintptr_t start, lv_uintptr_t end)
-{
-    LV_UNUSED(start);
-    LV_UNUSED(end);
-    LV_MEM_INVALIDATE_DCACHE;
 }
 
 /**********************
